Silence a clang warning on loss of precision
When converting a long or long long to a double, there may be loss
of precision and clang >= 10 warns about this unless there is a cast.
Add casts to silence this warning - the code is designed to handle
precisely this loss of precision, so this is harmless.

From CheriBSD via FreeBSD

ok millert

(cherry picked from commit 1f947e30b76f38b245ad8eb0af8005910b8fb08e)