Since futex(2) can fail, it needs a full syscall stub.  This should fixes issues on powerpc64 and sparc64.
Note that this makes the lib/libpthread/errno test fail since that test
detects that we are touching errno now.  This will be addressed in a
future diff as it is not entirely clear whether the test is correct.

ok mpi@, bluhm@

(cherry picked from commit 6a27e4b9577a558952acbc386a84bf6ecd3bba82)