store provider ID in umb(4), and display it in ifconfig.  OK stsp deraadt
Previously only the provider's display name was used. The text used depends
on how the SIM is configured and not just on the network in use (for example,
an MVNO SIM on another network will often display the MVNO's name rather
than that of the underlying network).

I have a SIM that roams to any network in my country - whichever network
it roams onto, the display name is the same, so you can't tell which
network you're really using. By printing the provider ID (in GSM-land this
is MCC+MNC) it's easy to lookup and check this.

As the provider was printed on the ifconfig line also showing subscriber-id
and ICCID it was already a bit long, and adding the provider-id there is
a bit too much, so move it to the output line showing APN, now looking like

:        subscriber-id 2400xxxxxxxxxxx ICC-id 8946203xxxxxxxxxxxxx
:        device EM7455 IMEI 01458xxxxxxxxxx firmware SWI9X30C_02.24.0
:        APN key provider Tele2 IoT provider-id 23420

(cherry picked from commit 5a5baf73d63c950ad2fd5afa5cfd410472f6f43e)