Make dhcpleased(8) always configure provided routes, regardless of whether the address we received in our lease is already configured.
In the case I observed, no default route was added to the routing table
even though the server provided both an address and a route option.
As it happened the leased address was already configured on the interface.
This should not prevent routing table updates, but it did.

ok florian

(cherry picked from commit 665300f1de7a790196e841021eeefbcdbb8f6d4b)