Add checks when reading VRPs with a maxLength via RTR
Ensure the maxLength is greater than or equal to the length of
the accompanying prefix, and less than or equal to the length
(in bits) of an IP address in the address family (32 for IPv4
and 128 for IPv6). The same checks are applied when reading VRPs
from configuration on disk.

OK claudio@

(cherry picked from commit 1eccb425429151dba6183280e853fbf072e4f528)