Back out last commit.
deraadt@ pointed out that the operation is correct.
It attaches, but by then it is being pulled.  The dying flag is set to
avoid half the kinds of failure conditions.  The other half of the failure
conditions are detect on-the-spot.

And it seems when compiled for debug, this on-the-spot check which notices
the device isn't there, has a printf.

(cherry picked from commit ca35d09358908198676e2c1492fa45e722582310)